+++ src/store/initialState.js
@@ -1,4 +1,5 @@
 import uuid from 'uuid';
+import { format } from 'date-fns';
 import CATEGORIES from '../utils/categories';
 
 const books = [
@@ -6,21 +7,21 @@ const books = [
     id: uuid.v4(),
     name: 'keep Me Calm',
     author: 'Dr, Dawuni Manzi',
-    year: Date.now(),
+    year: format(Date.now(), 'yyy MMM'),
     category: CATEGORIES[3],
   },
   {
     id: uuid.v4(),
     name: 'One Piece The Strongest',
     author: 'Erebor',
-    year: Date.now(),
+    year: format(Date.now(), 'yyy MMM'),
     category: CATEGORIES[0],
   },
   {
     id: '212121h2jhj',
     name: 'One Piece The Strongest ||',
     author: 'Erebor',
-    year: Date.now(),
+    year: format(Date.now(), 'yyy MMM'),
     category: CATEGORIES[0],
   },
